i run to you at sunrise when the dark has lost his grep

+++ he_who_remains.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   he_who_remains.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: brouane <brouane@student.1337.ma>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2026/04/29 12:33:44 by brouane           #+#    #+#             */
+/*   Updated: 2026/04/29 12:33:45 by brouane          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "codexion.h"
 
 int bye_bye()
@@ -15,8 +27,8 @@ void log_action(t_code_sim *code_sim, char *action)
 {
     if (!code_sim->sim->is_finished)
     {
-        unsigned long long timestamp = get_time_ms() - code_sim->sim->start_time;
-        // printf("get_time_ms() %lld - code_sim->coder->sim->start_time %lld\n", get_time_ms(), code_sim->sim->start_time);
-        printf("%lld %d %s\n", timestamp, code_sim->coder->coder_id, action);
+        unsigned long timestamp = get_time_ms() - code_sim->sim->start_time;
+        // printf("get_time_ms() %ld - code_sim->coder->sim->start_time %ld\n", get_time_ms(), code_sim->sim->start_time);
+        printf("%ld %d %s\n", timestamp, code_sim->coder->coder_id, action);
     }
 }
